# [Inductor][1/n]Split cat customization