--- a/torch/_inductor/fx_passes/split_cat.py
+++ b/torch/_inductor/fx_passes/split_cat.py
@@ -27,16 +27,7 @@ from ..pattern_matcher import (
     RepeatedExpr,
 )
 from .group_batch_fusion import is_node_meta_valid
-from .pre_grad import (
-    merge_getitem_cat_pass,
-    merge_splits_pass,
-    merge_stack_tahn_unbind_pass,
-    mutate_cat_pass,
-    normalization_pass,
-    remove_split_with_size_one_pass,
-    split_cat_pass,
-    unbind_stack_pass,
-)
+from .pre_grad import PRE_GRAD_PATTERNS
 
 log = logging.getLogger(__name__)
 
@@ -157,12 +148,12 @@ def normalize_split_base(
 
 @register_graph_pattern(
     CallFunctionVarArgs(torch.split, users=MULTIPLE),
-    pass_dict=normalization_pass,
+    pass_dict=PRE_GRAD_PATTERNS["normalization_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 @register_graph_pattern(
     CallMethodVarArgs("split", users=MULTIPLE),
-    pass_dict=normalization_pass,
+    pass_dict=PRE_GRAD_PATTERNS["normalization_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 def normalize_split_default(match: Match, *args, **kwargs):
@@ -171,12 +162,12 @@ def normalize_split_default(match: Match, *args, **kwargs):
 
 @register_graph_pattern(
     CallFunctionVarArgs(torch.split, users=MULTIPLE),
-    pass_dict=remove_split_with_size_one_pass,
+    pass_dict=PRE_GRAD_PATTERNS["remove_split_with_size_one_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 @register_graph_pattern(
     CallMethodVarArgs("split", users=MULTIPLE),
-    pass_dict=remove_split_with_size_one_pass,
+    pass_dict=PRE_GRAD_PATTERNS["remove_split_with_size_one_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 def remove_split_with_size_one(match: Match, *args, **kwargs):
@@ -211,12 +202,12 @@ def remove_split_with_size_one(match: Match, *args, **kwargs):
 
 @register_graph_pattern(
     CallFunctionVarArgs(torch.unbind, users=MULTIPLE),
-    pass_dict=normalization_pass,
+    pass_dict=PRE_GRAD_PATTERNS["normalization_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 @register_graph_pattern(
     CallMethodVarArgs("unbind", users=MULTIPLE),
-    pass_dict=normalization_pass,
+    pass_dict=PRE_GRAD_PATTERNS["normalization_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 def normalize_unbind_default(match: Match, *args, **kwargs):
@@ -253,7 +244,7 @@ def normalize_unbind_default(match: Match, *args, **kwargs):
 
 @register_graph_pattern(
     CallFunctionVarArgs(torch.cat, users=MULTIPLE),
-    pass_dict=normalization_pass,
+    pass_dict=PRE_GRAD_PATTERNS["normalization_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 def normalize_cat_default(match: Match, *args, **kwargs):
@@ -315,7 +306,7 @@ def normalize_cat_default(match: Match, *args, **kwargs):
 
 @register_graph_pattern(
     CallFunctionVarArgs(torch.stack, users=MULTIPLE),
-    pass_dict=normalization_pass,
+    pass_dict=PRE_GRAD_PATTERNS["normalization_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 def normalize_stack_default(match: Match, *args, **kwargs):
@@ -361,7 +352,7 @@ def find_next_users(split_node: torch.fx.Node) -> List[torch.fx.Node]:
 
 @register_graph_pattern(
     CallMethodVarArgs("squeeze", users=MULTIPLE),
-    pass_dict=normalization_pass,
+    pass_dict=PRE_GRAD_PATTERNS["normalization_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 def normalize_squeeze_default(match: Match, *args, **kwargs):
@@ -444,7 +435,7 @@ class TorchSplit(CallFunction):
         ),
         KeywordArg("next_split_sections"),
     ),
-    pass_dict=merge_splits_pass,
+    pass_dict=PRE_GRAD_PATTERNS["merge_splits_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 def merge_splits(
@@ -1050,7 +1041,7 @@ class GetItem(CallFunction):
             _users=MULTIPLE,
         ),
     ),
-    pass_dict=split_cat_pass,
+    pass_dict=PRE_GRAD_PATTERNS["split_cat_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 @register_graph_pattern(
@@ -1068,7 +1059,7 @@ class GetItem(CallFunction):
             _users=MULTIPLE,
         )
     ),
-    pass_dict=split_cat_pass,
+    pass_dict=PRE_GRAD_PATTERNS["split_cat_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 def merge_split_squeeze(
@@ -1122,21 +1113,21 @@ getitem_unbind = ListOf(
 
 @register_graph_pattern(
     CallFunction([torch.stack, torch.cat], getitem_unbind, Ignored(), _users=MULTIPLE),
-    pass_dict=unbind_stack_pass,
+    pass_dict=PRE_GRAD_PATTERNS["unbind_stack_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 @register_graph_pattern(
     CallFunction(
         [torch.stack, torch.cat], getitem_unbind, dim=Ignored(), _users=MULTIPLE
     ),
-    pass_dict=unbind_stack_pass,
+    pass_dict=PRE_GRAD_PATTERNS["unbind_stack_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 @register_graph_pattern(
     CallFunction(
         [torch.stack, torch.cat], tensors=getitem_unbind, dim=Ignored(), _users=MULTIPLE
     ),
-    pass_dict=unbind_stack_pass,
+    pass_dict=PRE_GRAD_PATTERNS["unbind_stack_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 def merge_unbind_stack(match: Match, unbind_input: torch.fx.Node, dim: int):
@@ -1165,7 +1156,7 @@ getitem_split = ListOf(
         dim=Ignored(),
         _users=MULTIPLE,
     ),
-    pass_dict=split_cat_pass,
+    pass_dict=PRE_GRAD_PATTERNS["split_cat_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 @register_graph_pattern(
@@ -1175,7 +1166,7 @@ getitem_split = ListOf(
         dim=Ignored(),
         _users=MULTIPLE,
     ),
-    pass_dict=split_cat_pass,
+    pass_dict=PRE_GRAD_PATTERNS["split_cat_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 @register_graph_pattern(
@@ -1185,7 +1176,7 @@ getitem_split = ListOf(
         Ignored(),
         _users=MULTIPLE,
     ),
-    pass_dict=split_cat_pass,
+    pass_dict=PRE_GRAD_PATTERNS["split_cat_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 def simplify_split_cat(match: Match, split_sections: List[int], dim: int):
@@ -1270,7 +1261,7 @@ def calculate_fused_tensor_size(split_node: torch.fx.Node, indices: List[int]) -
         dim=Ignored(),
         _users=MULTIPLE,
     ),
-    pass_dict=merge_getitem_cat_pass,
+    pass_dict=PRE_GRAD_PATTERNS["merge_getitem_cat_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 def merge_getitem_cat(match: Match, split_sections: List[int], dim: int):
@@ -1378,7 +1369,7 @@ def merge_getitem_cat(match: Match, split_sections: List[int], dim: int):
         dim=Ignored(),
         _users=MULTIPLE,
     ),
-    pass_dict=mutate_cat_pass,
+    pass_dict=PRE_GRAD_PATTERNS["mutate_cat_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 def mutate_cat_node(match: Match, split_sections: List[int], dim: int):
@@ -1474,7 +1465,7 @@ def mutate_cat_node(match: Match, split_sections: List[int], dim: int):
             dim=Ignored(),
         ),
     ),
-    pass_dict=merge_getitem_cat_pass,
+    pass_dict=PRE_GRAD_PATTERNS["merge_getitem_cat_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 @register_graph_pattern(
@@ -1486,7 +1477,7 @@ def mutate_cat_node(match: Match, split_sections: List[int], dim: int):
             dim=Ignored(),
         ),
     ),
-    pass_dict=merge_getitem_cat_pass,
+    pass_dict=PRE_GRAD_PATTERNS["merge_getitem_cat_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 @register_graph_pattern(
@@ -1498,7 +1489,7 @@ def mutate_cat_node(match: Match, split_sections: List[int], dim: int):
             Ignored(),
         ),
     ),
-    pass_dict=merge_stack_tahn_unbind_pass,
+    pass_dict=PRE_GRAD_PATTERNS["merge_stack_tahn_unbind_pass"],
     extra_check=config_flag("split_cat_fx_passes"),
 )
 def merge_stack_tahn_unbind(match: Match, split_sections: List[int], dim: int):